backport 19968 (quick damlc feedback for upgrade script tests)
